backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
enablebanking-sdk
python-dotenv
